# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Running Man API
  description: |
    REST API for querying logs and managing processes in Running Man.
    
    Running Man is a dev observability tool that aggregates logs from multiple sources
    (processes and Docker containers) and provides a unified API for querying them.
    
    **Security Note**: This API is designed for local development use only and runs on localhost
    without authentication. It should not be exposed to public networks.
  version: 1.0.0
  license:
    name: MIT
  contact:
    name: Running Man
    url: https://github.com/elbeanio/the_running_man
servers:
  - url: http://localhost
    description: |
      Served on a Unix socket, not a TCP port: .running-man/api.sock inside the
      project directory (mode 0600). The host in the URL is ignored -- reach it
      with curl's --unix-socket, or an equivalent transport. No authentication,
      because the filesystem performs it.
...
paths:
...
  /logs:
    get:
      summary: Query logs with filters
      description: |
        Retrieve log entries with optional filtering by time, level, source, and content.
        Supports glob patterns for source filtering and time-based queries.
      tags:
        - logs
      parameters:
        - name: since
          in: query
          description: Time duration to filter logs (e.g., "30s", "5m", "1h", "2h30m")
          required: false
          schema:
            type: string
            example: "30s"
        - name: level
          in: query
          description: Comma-separated log levels to filter (debug, info, warn, error)
          required: false
          schema:
            type: string
            example: "error,warn"
        - name: source
          in: query
          description: Comma-separated source names or glob patterns (e.g., "my-api", "postgres-*")
          required: false
          schema:
            type: string
            example: "my-api"
        - name: exclude
          in: query
          description: Comma-separated source exclusion patterns (supports globs)
          required: false
          schema:
            type: string
            example: "test-*"
        - name: contains
          in: query
          description: Filter logs containing this text in the message
          required: false
          schema:
            type: string
            example: "error connecting"
        - name: limit
          in: query
          description: |
            Maximum number of entries to return, keeping the MOST RECENT matches.
            Defaults to 1000. Use `limit=0` for no limit. Negative values are rejected.
          required: false
          schema:
            type: integer
            minimum: 0
            default: 1000
            example: 100
      responses:
        '200':
          description: List of log entries matching the filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
              example:
                logs:
                  - timestamp: "2026-02-09T23:15:30Z"
                    level: "error"
                    source: "my-api"
                    message: "Failed to connect to database"
                    raw: "[2026-02-09 23:15:30] ERROR: Failed to connect to database"
                    stacktrace: ""
                  - timestamp: "2026-02-09T23:15:32Z"
                    level: "warn"
                    source: "my-api"
                    message: "Retrying connection in 5s"
                    raw: "[2026-02-09 23:15:32] WARN: Retrying connection in 5s"
                    stacktrace: ""
                count: 2
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid since parameter: invalid duration format: 30x (expected: 30s, 5m, 1h)"
  /errors:
    get:
      summary: Get error logs only
      description: Retrieve only error-level log entries, optionally filtered by time
      tags:
        - logs
      parameters:
        - name: since
          in: query
          description: Time duration to filter errors (e.g., "30s", "5m", "1h")
          required: false
          schema:
            type: string
            example: "1h"
        - name: limit
          in: query
          description: |
            Maximum number of entries to return, keeping the MOST RECENT matches.
            Defaults to 1000. Use `limit=0` for no limit.
          required: false
          schema:
            type: integer
            minimum: 0
            default: 1000
            example: 50
      responses:
        '200':
          description: List of error log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorsResponse'
              example:
                errors:
                  - timestamp: "2026-02-09T23:15:30Z"
                    level: "error"
                    source: "my-api"
                    message: "Failed to connect to database"
                    raw: "[2026-02-09 23:15:30] ERROR: Failed to connect to database"
                    stacktrace: "Traceback (most recent call last):\n  File \"app.py\", line 42, in connect\n    conn = psycopg2.connect(dsn)\npsycopg2.OperationalError: could not connect to server"
                count: 1
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /processes:
    get:
      summary: List all managed processes
      description: Returns information about all processes managed by Running Man
      tags:
        - processes
      responses:
        '200':
          description: List of managed processes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessesResponse'
              example:
                processes:
                  - name: "my-api"
                    command: "python server.py"
                    pid: 12345
                    status: "running"
                    exit_code: -1
                    start_time: "2026-02-09T22:00:00Z"
                  - name: "worker"
                    command: "npm run worker"
                    pid: 12346
                    status: "running"
                    exit_code: -1
                    start_time: "2026-02-09T22:00:05Z"
                count: 2
        '503':
          description: Process manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Process manager not available"
  /processes/{name}:
    get:
      summary: Get process details
      description: Returns detailed information about a specific managed process
      tags:
        - processes
      parameters:
        - name: name
          in: path
          description: Name of the process
          required: true
          schema:
            type: string
            example: "my-api"
      responses:
        '200':
          description: Process details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInfo'
              example:
                name: "my-api"
                command: "python server.py"
                pid: 12345
                status: "running"
                exit_code: -1
                start_time: "2026-02-09T22:00:00Z"
        '400':
          description: Invalid process name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid process name"
        '404':
          description: Process not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Process not found: my-api"
        '503':
          description: Process manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /processes/{name}/restart:
    post:
      summary: Restart a process
      description: |
        Stops and restarts the specified managed process.

        Reaching the socket at all requires permission to open a file in the project
        directory, so there is no additional caller check. The loopback-only 403 this
        endpoint used to return went away with the TCP port.
      tags:
        - processes
      parameters:
        - name: name
          in: path
          description: Name of the process to restart
          required: true
          schema:
            type: string
            example: "my-api"
      responses:
        '200':
          description: Process restart initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  process:
                    $ref: '#/components/schemas/ProcessInfo'
              example:
                message: "Process restarted successfully"
                process:
                  name: "my-api"
                  command: "python server.py"
                  pid: 12347
                  status: "running"
                  exit_code: -1
                  start_time: "2026-02-09T23:25:00Z"
        '400':
          description: Invalid process name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Process not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to restart process
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to restart process: exit status 1"
        '503':
          description: Process manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /processes/stop-all:
    post:
      summary: Stop all managed processes
      description: |
        Gracefully stops all processes managed by Running Man.

        Reaching the socket at all requires permission to open a file in the project
        directory, so there is no additional caller check. The loopback-only 403 this
        endpoint used to return went away with the TCP port.
      tags:
        - processes
      responses:
        '200':
          description: All processes stopped successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  stopped_count:
                    type: integer
              example:
                message: "All processes stopped"
                stopped_count: 2
        '503':
          description: Process manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    LogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        count:
          type: integer
          description: Number of log entries returned
        limit:
          type: integer
          description: The limit applied to this query (0 means no limit)
      required:
        - logs
        - count
    ErrorsResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        count:
          type: integer
          description: Number of error entries returned
      required:
        - errors
        - count
...
    ProcessesResponse:
      type: object
      properties:
        processes:
          type: array
          items:
            $ref: '#/components/schemas/ProcessInfo'
        count:
          type: integer
          description: Number of processes
        dependencies:
          type: array
          description: |
            The Compose services that processes depend on. Empty when no process
            depends on one.
          items:
            $ref: '#/components/schemas/DependencyInfo'
      required:
        - processes
        - count
        - dependencies
    DependencyInfo:
      type: object
      description: A Compose service that a process depends on
      properties:
        name:
          type: string
          description: The Compose service name
        state:
          type: string
          enum: [pending, checking, ready, failed]
          description: |
            `pending` until something waits on it, `checking` while its healthcheck is
            being checked, then `ready` or `failed`. Running Man does not start Compose
            services, so a service that has been up for days is `checking` too, briefly.
        detail:
          type: string
          description: Why it failed. Absent otherwise.
        sources:
          type: array
          items:
            type: string
          description: The log sources of the service's containers, for `/logs?source=`.
      required:
        - name
        - state
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error